feat: trie with exact search working

+++ fileModule/fileModule.h
@@ -2,5 +2,5 @@
 
 namespace fileModule
 {
-  void processNodesToLabel(trie::TrieNode *trieRoot, void (*cb)(trie::TrieNode *trieRoot, std::string word, unsigned int nodeId));
+  void processNodesToLabel(trie::TrieNode *trieRoot, void (*cb)(trie::TrieNode *trieRoot, std::string word, unsigned long nodeId));
 }
